Correções no fluxo de análise

+++ src/main/java/com/fishcount/api/repository/ParametroTemperaturaRepository.java
@@ -5,10 +5,12 @@ import com.fishcount.common.model.entity.ParametroTemperatura;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
+import java.math.BigDecimal;
+
 public interface ParametroTemperaturaRepository
         extends JpaRepository<ParametroTemperatura, Integer>, JpaSpecificationExecutor<ParametroTemperatura>, CustomParametroTemperaturaRepository {
 
     @Override
-    ParametroTemperatura findByTemperatura(Integer temperatura);
+    ParametroTemperatura findByTemperatura(BigDecimal temperatura);
 
 }
