+++ src/main/java/com/fishcount/api/repository/custom/CustomParametroTemperaturaRepository.java
@@ -2,8 +2,10 @@ package com.fishcount.api.repository.custom;
 
 import com.fishcount.common.model.entity.ParametroTemperatura;
 
+import java.math.BigDecimal;
+
 public interface CustomParametroTemperaturaRepository {
 
-    ParametroTemperatura findByTemperatura(Integer temperatura);
+    ParametroTemperatura findByTemperatura(BigDecimal temperatura);
 
 }
